requirements.txt
customtkinter>=5.0.0
pillow>=9.0.0
requests>=2.25.0
